Refactor Baby-Actions items (#13)
* part 1

* part 2

* part 3

* part 4

* part 5

* part 6

---------

+++ src/app/baby-actions/baby-actions-info/baby-actions-info-item/baby-actions-info-item.component.ts
@@ -1,29 +1,42 @@
-import { Component, ElementRef, Input, ViewChild } from '@angular/core';
+import {
+  Component,
+  ElementRef,
+  Input,
+  Output,
+  EventEmitter,
+  ViewChild,
+} from '@angular/core';
 import { BabyActionDataModel } from '../../../models/baby-action-data.model';
 import { BabyActionsDataService } from '../../../services/baby-actions-data.service';
 
 @Component({
   selector: 'app-baby-actions-info-item',
   templateUrl: './baby-actions-info-item.component.html',
-  styleUrl: './baby-actions-info-item.component.css',
+  styleUrls: ['./baby-actions-info-item.component.css'],
 })
 export class BabyActionsInfoItemComponent {
   @Input() babyActionData: BabyActionDataModel | undefined;
   @Input() index: number | undefined;
+  @Input() filterMode: boolean = false; // Receive shared filterMode
+  @Output() filter = new EventEmitter<any>();
+  @Output() unfilter = new EventEmitter<void>();
   @ViewChild('descriptionTextarea', { static: true })
   descriptionTextareaRef: ElementRef;
   editMode = false;
 
   constructor(private babyActionsDataService: BabyActionsDataService) {}
 
-  onStartEditMode(event: MouseEvent) {
-    event.preventDefault();
+  onStartEditMode(event?: MouseEvent) {
+    if (event) event.preventDefault();
     this.editMode = true;
   }
+
   onEditSubmit(newDescription: string) {
-    this.babyActionData.description = newDescription;
-    this.editMode = false;
-    this.babyActionsDataService.updatedBabyAction(this.babyActionData);
+    if (this.babyActionData) {
+      this.babyActionData.description = newDescription;
+      this.editMode = false;
+      this.babyActionsDataService.updatedBabyAction(this.babyActionData);
+    }
   }
 
   onCancelEdit() {
@@ -31,6 +44,18 @@ export class BabyActionsInfoItemComponent {
   }
 
   onDelete() {
-    this.babyActionsDataService.deleteBabyAction(this.babyActionData);
+    if (this.babyActionData) {
+      this.babyActionsDataService.deleteBabyAction(this.babyActionData);
+    }
+  }
+
+  onFilter() {
+    if (this.babyActionData) {
+      this.filter.emit(this.babyActionData.category);
+    }
+  }
+
+  onUnfilter() {
+    this.unfilter.emit();
   }
 }
